Export all check and contact structures
As I was preparing for integration testing I realized that several
structs that I didn't think I would need to export would probably
require exporting to ensure a flexible development workflow, at the very
least for testing.

As part of the export I have pluralized the packages to try and reduce
stutter. This also matches the URI conventions on Pingdom, for what it's
worth.